pm: enable reset prep IRQ
In the SoC warm reboot process, sometimes there are
instances where the ISH firmware does not receive the
warm reboot request, leading to a failure to reboot ISH
firmware leading to ISHTP driver communication failure with
ISH firmware. This issue is observred when there is an IPC
request pending at the ISH and at the time a warm reboot is
initiated causing IPM driver timeout and IPC communication
failure.

change provided will re-enable the RESET PREP IRQ in the
sedi_pm_enter_power_state API. This change helps the ISH PM
task to handle the RESET PREP ISR routine whenever there is a
RESET PREP IRQ receieved. This ISR is responsible for handling
the ISH reset, thereby mitigating the issue.

Change-Id: I454bd75257cdfc1c8f37e582231f6fe7f5e8f91d